requirements/conda.txt
cloudpickle
pygments >=2.0
qtconsole >=4.5.0
nbconvert
sphinx
pylint
psutil
qtawesome >=0.5.2
qtpy >=1.5.0
pickleshare
pyzmq
chardet >=2.0.0
numpydoc
pyqt <5.10
keyring
spyder-kernels >=1.2
python-language-server >=0.19
qdarkstyle >=2.6.4
atomicwrites
diff-match-patch
pexpect
watchdog
